requirements.txt
gunicorn
django-heroku
dj-database-url
django-allauth==0.61.1
django-bootstrap-v5
django-storages
boto3
python-dotenv
django-background-tasks